Combine media_player tests into one file
